changing quantity field from int to long

+++ src/main/java/com/br/felipe/ecommerce/models/Item.java
@@ -25,7 +25,7 @@ public class Item extends RepresentationModel<Item> implements Serializable{
 	
 	private String description;
 	
-	private int qtd;
+	private long qtd;
 
 	public Item() {
